+++ bin/cli.js
@@ -2,16 +2,13 @@
 'use strict';
 
 const process = require('process');
-const path = require('path');
-const fs = require('mz/fs');
 const program = require('commander');
 const rc = require('rc')('madge');
-const readPackage = require('read-package-json');
 const version = require('../package.json').version;
 
 program
 	.version(version)
-	.usage('[options] [file|dir]')
+	.usage('[options] <src...>')
 	.option('--basedir <path>', 'base directory for resolving paths')
 	.option('--list', 'show dependency list (default)')
 	.option('--summary', 'show dependency count summary')
@@ -22,11 +19,18 @@ program
 	.option('--image <file>', 'write graph to file as an image')
 	.option('--layout <name>', 'layout engine to use for graph (dot/neato/fdp/sfdp/twopi/circo)')
 	.option('--dot', 'show graph using the DOT language')
+	.option('--extensions <list>', 'comma separated string of valid file extensions')
+	.option('--show-extension', 'include file extension in module name', false)
 	.option('--require-config <file>', 'path to RequireJS config')
 	.option('--webpack-config <file>', 'path to webpack config')
 	.option('--no-color', 'disable color in output and image', false)
 	.option('--debug', 'turn on debug output', false)
 	.parse(process.argv);
+
+if (!program.args.length) {
+	console.log(program.helpInformation());
+	process.exit(1);
+}
 
 if (program.debug) {
 	process.env.DEBUG = '*';
@@ -39,7 +43,6 @@ if (!program.color) {
 const log = require('../lib/log');
 const output = require('../lib/output');
 const madge = require('../lib/api');
-const target = program.args[0] || process.cwd();
 const config = Object.assign({}, rc);
 
 delete config._;
@@ -47,7 +50,7 @@ delete config.config;
 delete config.configs;
 
 if (rc.config) {
-	log('using runtime configuration from %s', rc.config);
+	log('using runtime config %s', rc.config);
 }
 
 ['layout', 'requireConfig', 'webpackConfig'].forEach((option) => {
@@ -64,6 +67,14 @@ if (program.exclude) {
 	config.excludeRegExp = [program.exclude];
 }
 
+if (program.extensions) {
+	config.fileExtensions = program.extensions.split(',').map((s) => s.trim());
+}
+
+if (program.showExtension) {
+	config.showFileExtension = true;
+}
+
 if (!program.color) {
 	config.backgroundColor = '#ffffff';
 	config.nodeColor = '#00000';
@@ -72,37 +83,7 @@ if (!program.color) {
 	config.edgeColor = '#757575';
 }
 
-fs
-	.stat(target)
-	.then((stats) => {
-		if (stats.isFile()) {
-			return program.args[0];
-		}
-
-		const pkgPath = path.join(target, 'package.json');
-
-		return new Promise((resolve, reject) => {
-			readPackage(pkgPath, (err, pkg) => {
-				if (err) {
-					reject('Could not read ' + pkgPath + '. Choose another directory or specify file.');
-					return;
-				}
-
-				config.baseDir = target;
-
-				if (pkg.bin) {
-					log('extracted file path from "bin" entry in ' + pkgPath);
-					resolve(path.join(target, pkg.bin[Object.keys(pkg.bin)[0]]));
-				} else if (pkg.main) {
-					log('extracted file path from "main" entry in ' + pkgPath);
-					resolve(path.join(target, pkg.main));
-				} else {
-					reject('Could not get file to scan by reading package.json. Update package.json or specify a file.');
-				}
-			});
-		});
-	})
-	.then((filePath) => madge(filePath, config))
+madge(program.args, config)
 	.then((res) => {
 		if (program.summary) {
 			return output.summary(res.obj(), {