+++ test/api.js
@@ -9,26 +9,64 @@ const madge = require('../lib/api');
 require('should');
 
 describe('Madge', () => {
-	it('throws error on missing filename argument', () => {
+	it('throws error on missing path argument', () => {
 		(() => {
 			madge();
-		}).should.throw('Filename argument is missing');
+		}).should.throw('path argument not provided');
 	});
 
 	it('returns a Promise', () => {
 		madge(__dirname + '/files/cjs/a.js').should.be.Promise(); // eslint-disable-line new-cap
 	});
 
-	it('throws error if filename argument is not a file', (done) => {
-		madge(__dirname + '/files').catch((err) => {
-			err.message.should.match(/is not a file/);
+	it('throws error if file or directory does not exists', (done) => {
+		madge(__dirname + '/missing.js').catch((err) => {
+			err.message.should.match(/no such file or directory/);
 			done();
 		}).catch(done);
 	});
 
-	it('throws error if file does not exists', (done) => {
-		madge(__dirname + '/missing.js').catch((err) => {
-			err.message.should.match(/does not exists/);
+	it('takes single file as path', (done) => {
+		madge(__dirname + '/files/cjs/a.js').then((res) => {
+			res.obj().should.eql({
+				'a': ['b', 'c'],
+				'b': ['c'],
+				'c': []
+			});
+			done();
+		}).catch(done);
+	});
+
+	it('takes an array of files as path and combines the result', (done) => {
+		madge([__dirname + '/files/cjs/a.js', __dirname + '/files/cjs/normal/d.js']).then((res) => {
+			res.obj().should.eql({
+				'a': ['b', 'c'],
+				'b': ['c'],
+				'c': [],
+				'normal/d': []
+			});
+			done();
+		}).catch(done);
+	});
+
+	it('take a single directory as path and find files in it', (done) => {
+		madge(__dirname + '/files/cjs/normal').then((res) => {
+			res.obj().should.eql({
+				'a': ['sub/b'],
+				'd': [],
+				'sub/b': ['sub/c'],
+				'sub/c': ['d']
+			});
+			done();
+		}).catch(done);
+	});
+
+	it('takes an array of directories as path and compute the basedir correctly', (done) => {
+		madge([__dirname + '/files/cjs/multibase/1', __dirname + '/files/cjs/multibase/2']).then((res) => {
+			res.obj().should.eql({
+				'1/a': [],
+				'2/b': []
+			});
 			done();
 		}).catch(done);
 	});