+++ CHANGELOG.md
@@ -20,8 +20,7 @@ So I decided it was high time for version 1.0 to be released and take the opport
 * Renamed many of the settings in the [config](README.md#configuration)
 * Option `--json` should now be used instead of `--output json`
 * The generation of the dependency tree is now delegated to the external module [dependency-tree](https://github.com/mrjoelkemp/node-dependency-tree)
-* A single file will now be used as an entry instead of scanning entire folder(s)
-* Dependencies will now be extracted recursively from the single file
+* Dependencies will now be extracted recursively
 * NPM installed dependencies are now excluded by default
 * Node.js core modules are now excluded
 * The [API](README.md#api) is now using promises
@@ -33,7 +32,6 @@ So I decided it was high time for version 1.0 to be released and take the opport
 * Option `--read`
 * Option `--find-nested-dependencies`
 * Option `--paths`
-* Option `--extensions`
 * Option `--config`
 * Option `--output`
 * Option `--break-on-error`